requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
wandb
streamlit